requirements.txt
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
openai>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
